Sistemas con templates

+++ src/cmp/physicscmp2.hpp
@@ -2,8 +2,8 @@
 #include <cstdint>
 
 struct PhysicsCmp2 {
-    int  x{},  y{},  z{};
-    int vx{}, vy{}, vz{};
+    float  x{},  y{},  z{};
+    float vx{}, vy{}, vz{};
 
     static constexpr uint8_t id {0};
 };
